requirement.txt
django
pymongo==3.12.1
dnspython
djongo
pytz
djangorestframework
django-cors-headers
djangorestframework-simplejwt
